server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
firebase-admin==6.2.0
openai-whisper==20231117
yt-dlp==2023.11.16
numpy==1.26.2
torch==2.1.1
httpx==0.25.2
python-dotenv==1.0.0
pydantic==2.5.2 